Worked on update methods for all files
