+++ lib/invoice_items_repo.rb
@@ -5,10 +5,14 @@ require 'bigdecimal/util'
 require_relative './module'
 require_relative './invoice_items'
 require_relative './sales_engine'
+require_relative './central_repo'
 
-class InvoiceItemRepo
-  include Methods
+class InvoiceItemRepo < CentralRepo
   attr_reader :collections
+
+  def initialize(data, engine)
+    super
+  end
 
   def populate_collection
     invoice_items = Hash.new{|h, k| h[k] = [] }
